Moved tests so they are addressing a single issue now. We need to test the class generators at some point.
